Coursework
Successfully completed Uplimit’s SQL Crash Course, which was a 2 week crash course in the foundations of SQL including filtering and sorting data, aggregations, complex filtering logic and built-in date/string predicates, and the GROUP BY and HAVING clause. I was able to complete projects where I queried data including an analysis of pricing and revenue for Airbnb listings in Sydney, Australia.